Add RuntimeChoice to many-types
