+++ src/many-types/Cargo.toml
@@ -17,6 +17,7 @@ num-traits = "0.2.14"
 num-bigint = "0.4.3"
 proptest = { version = "1.0.0", optional = true }
 serde = "1.0.139"
+tokio = "1.12"
 
 [dev-dependencies]
 many-types = { path = ".", features = ["proptest"] }
